zk-dump: add --unpaired options
By default, zk-dump will wait for a reply from the server
to be seen and then it'll pair that with the request that
it saw before.

This new option, --unpaired, prints reqs, reps & evts as soon
as they are seen. This loses the order, but can be helpful
to debug certain timeouts & other scenarios.

Note, note, note: we are still subject to the whatever ordering
pcap sees from the Kernel. So absolute ordering, as what's seen
through the wire, isn't assured (but it probably gets close enough).

Signed-off-by: Raul Gutierrez S <rgs@twitter.com>